backend/requirements.txt
pymongo
PyJWT
bcrypt
python-dotenv